tasker/backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
dateparser==1.2.0
python-dateutil==2.9.0.post0

